+++ api/reports.go.old
@@ -1,0 +1,36 @@
+// package api
+
+// Deprecated, but may be repurposed. Currently avoiding cyclical import.
+// import (
+// 	"bytes"
+// 	"encoding/json"
+// 	"fmt"
+// 	"net/http"
+// 	"net/url"
+
+// 	log "github.com/sirupsen/logrus"
+
+// 	"github.com/reliablyhq/cli/config"
+// 	"github.com/reliablyhq/cli/core/report"
+// )
+
+// func SendReport(client *Client, orgID string, r *report.Report) (string, error) {
+// 	path := fmt.Sprintf("orgs/%s/reports", orgID)
+
+// 	var body bytes.Buffer
+// 	if err := json.NewEncoder(&body).Encode(r); err != nil {
+// 		return "", fmt.Errorf("failed to serialize: %s", err)
+// 	}
+
+// 	type savedReport struct {
+// 		ID string `json:"id"`
+// 	}
+// 	var response *savedReport
+
+// 	err := client.REST(config.Hostname, http.MethodPost, path, &body, &response)
+// 	if err == nil {
+// 		log.Debugf("Report %s has been saved", response.ID)
+// 	}
+
+// 	return response.ID, err
+// }
